move classes from xml_upload.py to own files in knora.dsplib.models
