+++ knora/dsplib/models/xmlbitstream.py
@@ -1,0 +1,26 @@
+from lxml import etree
+
+
+class XMLBitstream:
+    """Represents a bitstream object (file) of a resource in the XML used for data import"""
+
+    _value: str
+    _permissions: str
+
+    def __init__(self, node: etree.Element) -> None:
+        self._value = node.text
+        self._permissions = node.get('permissions')
+
+    @property
+    def value(self) -> str:
+        """The file path of the bitstream object"""
+        return self._value
+
+    @property
+    def permissions(self) -> str:
+        """Reference to the set of permissions for the bitstream object"""
+        return self._permissions
+
+    def print(self) -> None:
+        """Prints the bitstream object and its attributes."""
+        print('   Bitstream file path: ' + str(self._value))
